feat: add logic to check for no students and refactor filter
