+++ frontend/src/pages/notifications.tsx
@@ -7,7 +7,7 @@ export default function Notifications() {
   return (
     <main>
       <div className="space-y-[20px]">
-        <div className="font-[alternate-gothic] text-4xl">Notifications</div>
+        <div className="font-[alternate-gothic] text-2xl lg:text-4xl">Notifications</div>
         <div className="font-[Poppins] text-[16px]">
           Review information of new account creations below to approve or deny them.{" "}
         </div>
